Fix calculator bug on entries and result

+++ src/main/kotlin/com/github/jairrab/calc/lib/controls/buttons/OperatorProcessor.kt
@@ -28,7 +28,7 @@ internal class OperatorProcessor(
             entriesManager.isReadyToClear() -> {
                 entriesManager.setReadyToClear(false)
                 entriesManager.clearEntries()
-                entriesManager.addEntry(entriesManager.getResult().toString())
+                entriesManager.addEntry(entriesManager.getResult().toDouble().toString())
                 entriesManager.addEntry(operator)
             }
             entriesManager.isLastEntryAnOperator() -> {
